Support to build on MinGW
Mingw is defining _WIN32. Thus, it uses strndup function which is written in toml.c. But it shows error because number of argument for malloc function argument is different from MSVC.
Mingw shows error strndup function does not found if I made mingw gcc does not use strndup for _WIN32
As the result, I added check code with __GNUC__ function