Prevent searching the entire string for every double-quote.
Issue #52: The logic to check for a triple s-quote within a double-quoted
string ran for every double-quoted string and traversed the entire
remainder of the string. For large TOML documents this dramatically
slowed down the parse.

This changeset fixes that by only searching for the triple s-quote
after the end of the double-quoted string is found.